refactor(bars): introduce {Vertical,Horizontal}BarSeries orientation parents
Class-hierarchy reorg — no behavior change. Lifts orientation-specific
concerns out of CoreColumnSeries / CoreRowSeries into two new abstract
intermediate classes:

  BarSeries<T,V,L>
  ├── VerticalBarSeries<T,V,L,TError>      (column-orientation parent)
  │   └── CoreColumnSeries<T,V,L,TError>
  └── HorizontalBarSeries<T,V,L,TError>    (row-orientation parent)
      └── CoreRowSeries<T,V,L,TError>

What moves to the orientation parents (identical or symmetric between
the two pre-refactor — separating them lifts the abstraction without
hiding the per-orientation differences):

  - SoftDeleteOrDisposePoint (column collapses Height->0 at pivot;
    row collapses Width->0)
  - SetDefaultPointTransitions (visual + error visual Animate;
    identical between the two but uses TErrorGeometry, so it lives
    one level down from BarSeries)
  - GetRequestedSecondaryOffset => 0.5f
  - GetProperties helper (vertical / horizontal SeriesProperties flags)
  - HorizontalBarSeries.GetBounds (the axis-swap rebuild — formerly
    a CoreRowSeries override)
  - DataPadding default (0,1) for vertical, (1,0) for horizontal

CoreColumnSeries shrinks from 408 LOC to 343 (only the per-frame
Invalidate orchestration remains, to be lifted next). CoreRowSeries
shrinks similarly. CoreStackedColumnSeries / CoreStackedRowSeries
inherit through unchanged.

Both 26 animation trajectory tests and 23 bar-family snapshot tests
still pass — the reorg is pure inheritance plumbing.
