+++ src/LiveChartsCore/HorizontalBarSeries.cs
@@ -1,0 +1,179 @@
+// The MIT License(MIT)
+//
+// Copyright(c) 2021 Alberto Rodriguez Orozco & LiveCharts Contributors
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+using System;
+using System.Collections.Generic;
+using LiveChartsCore.Drawing;
+using LiveChartsCore.Drawing.Segments;
+using LiveChartsCore.Kernel;
+using LiveChartsCore.Kernel.Sketches;
+using LiveChartsCore.Measure;
+
+namespace LiveChartsCore;
+
+/// <summary>
+/// Abstract base for bar-shaped series that grow horizontally (rows and their
+/// stacked / range variants). Owns orientation-specific concerns that are
+/// independent of the per-point sizing math: primary axis is X, bars enter from
+/// the pivot at zero width, the soft-delete collapse animates Width -> 0, and
+/// <see cref="GetBounds"/> swaps the X / Y axis roles when delegating to
+/// <see cref="Kernel.Providers.DataFactory.GetCartesianBounds"/>.
+/// </summary>
+/// <typeparam name="TModel">The model type.</typeparam>
+/// <typeparam name="TVisual">The visual type.</typeparam>
+/// <typeparam name="TLabel">The label type.</typeparam>
+/// <typeparam name="TErrorGeometry">The error-bar geometry type.</typeparam>
+public abstract class HorizontalBarSeries<TModel, TVisual, TLabel, TErrorGeometry>
+    : BarSeries<TModel, TVisual, TLabel>
+        where TVisual : BoundedDrawnGeometry, new()
+        where TLabel : BaseLabelGeometry, new()
+        where TErrorGeometry : BaseLineGeometry, new()
+{
+    /// <summary>
+    /// Initializes a new instance of the <see cref="HorizontalBarSeries{TModel, TVisual, TLabel, TErrorGeometry}"/> class.
+    /// </summary>
+    /// <param name="values">The values.</param>
+    /// <param name="isStacked">Whether the series is stacked.</param>
+    protected HorizontalBarSeries(IReadOnlyCollection<TModel>? values, bool isStacked = false)
+        : base(GetProperties(isStacked), values)
+    {
+        DataPadding = new LvcPoint(1, 0);
+    }
+
+    /// <inheritdoc cref="CartesianSeries{TModel, TVisual, TLabel}.GetRequestedSecondaryOffset"/>
+    protected override double GetRequestedSecondaryOffset() => 0.5f;
+
+    /// <inheritdoc cref="Series{TModel, TVisual, TLabel}.SetDefaultPointTransitions(ChartPoint)"/>
+    protected override void SetDefaultPointTransitions(ChartPoint chartPoint)
+    {
+        var chart = chartPoint.Context.Chart;
+        if (chartPoint.Context.Visual is not TVisual visual) throw new Exception("Unable to initialize the point instance.");
+
+        var animation = GetAnimation(chart.CoreChart);
+
+        visual.Animate(animation);
+
+        if (chartPoint.Context.AdditionalVisuals is not null)
+        {
+            var e = (ErrorVisual<TErrorGeometry>)chartPoint.Context.AdditionalVisuals;
+            e.YError.Animate(animation);
+            e.XError.Animate(animation);
+        }
+    }
+
+    /// <inheritdoc cref="CartesianSeries{TModel, TVisual, TLabel}.SoftDeleteOrDisposePoint(ChartPoint, Scaler, Scaler)"/>
+    protected internal override void SoftDeleteOrDisposePoint(ChartPoint point, Scaler primaryScale, Scaler secondaryScale)
+    {
+        var visual = (TVisual?)point.Context.Visual;
+        if (visual is null) return;
+        if (DataFactory is null) throw new Exception("Data provider not found");
+
+        var p = primaryScale.ToPixels(pivot);
+        var secondary = secondaryScale.ToPixels(point.Coordinate.SecondaryValue);
+
+        visual.X = p;
+        visual.Y = secondary - visual.Height * 0.5f;
+        visual.Width = 0;
+        visual.RemoveOnCompleted = true;
+
+        if (point.Context.AdditionalVisuals is not null)
+        {
+            var e = (ErrorVisual<TErrorGeometry>)point.Context.AdditionalVisuals;
+
+            e.YError.Y = secondary - visual.Height * 0.5f;
+            e.YError.Y1 = secondary - visual.Height * 0.5f;
+            e.YError.RemoveOnCompleted = true;
+
+            e.XError.X = p;
+            e.XError.X1 = p;
+            e.XError.RemoveOnCompleted = true;
+        }
+
+        DataFactory.DisposePoint(point);
+
+        var label = (TLabel?)point.Context.Label;
+        if (label is null) return;
+
+        label.TextSize = 1;
+        label.RemoveOnCompleted = true;
+    }
+
+    /// <inheritdoc cref="CartesianSeries{TModel, TVisual, TLabel}.GetBounds(Chart, ICartesianAxis, ICartesianAxis)"/>
+    public override SeriesBounds GetBounds(Chart chart, ICartesianAxis secondaryAxis, ICartesianAxis primaryAxis)
+    {
+        var rawBounds = DataFactory.GetCartesianBounds(chart, this, primaryAxis, secondaryAxis);
+        if (rawBounds.HasData) return rawBounds;
+
+        var rawBaseBounds = rawBounds.Bounds;
+
+        var tickPrimary = primaryAxis.GetTick(chart.ControlSize, rawBaseBounds.VisibleSecondaryBounds);
+        var tickSecondary = secondaryAxis.GetTick(chart.ControlSize, rawBaseBounds.VisiblePrimaryBounds);
+
+        var ts = tickSecondary.Value * DataPadding.X;
+        var tp = tickPrimary.Value * DataPadding.Y;
+
+        var rgs = GetRequestedGeometrySize();
+        var rso = GetRequestedSecondaryOffset();
+        var rpo = GetRequestedPrimaryOffset();
+
+        var dimensionalBounds = new DimensionalBounds
+        {
+            SecondaryBounds = new Bounds
+            {
+                Max = rawBaseBounds.PrimaryBounds.Max + rpo * secondaryAxis.UnitWidth,
+                Min = rawBaseBounds.PrimaryBounds.Min - rpo * secondaryAxis.UnitWidth,
+                MinDelta = rawBaseBounds.PrimaryBounds.MinDelta,
+                PaddingMax = ts,
+                PaddingMin = ts,
+                RequestedGeometrySize = rgs
+            },
+            PrimaryBounds = new Bounds
+            {
+                Max = rawBaseBounds.SecondaryBounds.Max + rso * primaryAxis.UnitWidth,
+                Min = rawBaseBounds.SecondaryBounds.Min - rso * primaryAxis.UnitWidth,
+                MinDelta = rawBaseBounds.SecondaryBounds.MinDelta,
+                PaddingMax = tp,
+                PaddingMin = tp,
+                RequestedGeometrySize = rgs
+            },
+            VisibleSecondaryBounds = new Bounds
+            {
+                Max = rawBaseBounds.VisiblePrimaryBounds.Max + rpo * secondaryAxis.UnitWidth,
+                Min = rawBaseBounds.VisiblePrimaryBounds.Min - rpo * secondaryAxis.UnitWidth
+            },
+            VisiblePrimaryBounds = new Bounds
+            {
+                Max = rawBaseBounds.VisibleSecondaryBounds.Max + rso * primaryAxis.UnitWidth,
+                Min = rawBaseBounds.VisibleSecondaryBounds.Min - rso * primaryAxis.UnitWidth,
+            },
+            TertiaryBounds = rawBaseBounds.TertiaryBounds,
+            VisibleTertiaryBounds = rawBaseBounds.VisibleTertiaryBounds
+        };
+
+        return new SeriesBounds(dimensionalBounds, false);
+    }
+
+    private static SeriesProperties GetProperties(bool isStacked) =>
+        SeriesProperties.Bar | SeriesProperties.PrimaryAxisHorizontalOrientation |
+        SeriesProperties.Solid | SeriesProperties.PrefersYStrategyTooltips |
+        (isStacked ? SeriesProperties.Stacked : 0);
+}
