+++ src/LiveChartsCore/VerticalBarSeries.cs
@@ -1,0 +1,121 @@
+// The MIT License(MIT)
+//
+// Copyright(c) 2021 Alberto Rodriguez Orozco & LiveCharts Contributors
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+using System;
+using System.Collections.Generic;
+using LiveChartsCore.Drawing;
+using LiveChartsCore.Drawing.Segments;
+using LiveChartsCore.Kernel;
+using LiveChartsCore.Measure;
+
+namespace LiveChartsCore;
+
+/// <summary>
+/// Abstract base for bar-shaped series that grow vertically (columns and their
+/// stacked / range variants). Owns orientation-specific concerns that are
+/// independent of the per-point sizing math: primary axis is Y, bars enter from
+/// the pivot at zero height, the soft-delete collapse animates Height -> 0.
+/// </summary>
+/// <typeparam name="TModel">The model type.</typeparam>
+/// <typeparam name="TVisual">The visual type.</typeparam>
+/// <typeparam name="TLabel">The label type.</typeparam>
+/// <typeparam name="TErrorGeometry">The error-bar geometry type.</typeparam>
+public abstract class VerticalBarSeries<TModel, TVisual, TLabel, TErrorGeometry>
+    : BarSeries<TModel, TVisual, TLabel>
+        where TVisual : BoundedDrawnGeometry, new()
+        where TLabel : BaseLabelGeometry, new()
+        where TErrorGeometry : BaseLineGeometry, new()
+{
+    /// <summary>
+    /// Initializes a new instance of the <see cref="VerticalBarSeries{TModel, TVisual, TLabel, TErrorGeometry}"/> class.
+    /// </summary>
+    /// <param name="values">The values.</param>
+    /// <param name="isStacked">Whether the series is stacked.</param>
+    protected VerticalBarSeries(IReadOnlyCollection<TModel>? values, bool isStacked = false)
+        : base(GetProperties(isStacked), values)
+    {
+        DataPadding = new LvcPoint(0, 1);
+    }
+
+    /// <inheritdoc cref="CartesianSeries{TModel, TVisual, TLabel}.GetRequestedSecondaryOffset"/>
+    protected override double GetRequestedSecondaryOffset() => 0.5f;
+
+    /// <inheritdoc cref="Series{TModel, TVisual, TLabel}.SetDefaultPointTransitions(ChartPoint)"/>
+    protected override void SetDefaultPointTransitions(ChartPoint chartPoint)
+    {
+        var chart = chartPoint.Context.Chart;
+        if (chartPoint.Context.Visual is not TVisual visual) throw new Exception("Unable to initialize the point instance.");
+
+        var animation = GetAnimation(chart.CoreChart);
+
+        visual.Animate(animation);
+
+        if (chartPoint.Context.AdditionalVisuals is not null)
+        {
+            var e = (ErrorVisual<TErrorGeometry>)chartPoint.Context.AdditionalVisuals;
+            e.YError.Animate(animation);
+            e.XError.Animate(animation);
+        }
+    }
+
+    /// <inheritdoc cref="CartesianSeries{TModel, TVisual, TLabel}.SoftDeleteOrDisposePoint(ChartPoint, Scaler, Scaler)"/>
+    protected internal override void SoftDeleteOrDisposePoint(ChartPoint point, Scaler primaryScale, Scaler secondaryScale)
+    {
+        var visual = (TVisual?)point.Context.Visual;
+        if (visual is null) return;
+        if (DataFactory is null) throw new Exception("Data provider not found");
+
+        var p = primaryScale.ToPixels(pivot);
+        var secondary = secondaryScale.ToPixels(point.Coordinate.SecondaryValue);
+
+        visual.X = secondary - visual.Width * 0.5f;
+        visual.Y = p;
+        visual.Height = 0;
+        visual.RemoveOnCompleted = true;
+
+        if (point.Context.AdditionalVisuals is not null)
+        {
+            var e = (ErrorVisual<TErrorGeometry>)point.Context.AdditionalVisuals;
+
+            e.YError.Y = p;
+            e.YError.Y1 = p;
+            e.YError.RemoveOnCompleted = true;
+
+            e.XError.X = secondary - visual.Width * 0.5f;
+            e.XError.X1 = secondary - visual.Width * 0.5f;
+            e.XError.RemoveOnCompleted = true;
+        }
+
+        DataFactory.DisposePoint(point);
+
+        var label = (TLabel?)point.Context.Label;
+        if (label is null) return;
+
+        label.TextSize = 1;
+        label.RemoveOnCompleted = true;
+    }
+
+    private static SeriesProperties GetProperties(bool isStacked) =>
+        SeriesProperties.Bar | SeriesProperties.PrimaryAxisVerticalOrientation |
+        SeriesProperties.Solid | SeriesProperties.PrefersXStrategyTooltips |
+        (isStacked ? SeriesProperties.Stacked : 0);
+}
